+++ src/swarm_rescue/spg_overlay/misc_data.py
@@ -8,5 +8,7 @@ class MiscData:
 
     def __init__(self,
                  size_area: Tuple[float, float] = None,
+                 number_drones: int = None
                  ):
         self.size_area = size_area
+        self.number_drones = number_drones